RichMenu系コンポーネントの分離
