+++ src/components/organisms/RichMenu/RichMenuContents/RichMenuContents.stories.tsx
@@ -13,5 +13,3 @@ export default {
 export const Default: StoryObj<typeof RichMenuContents> = {
   args: {},
 };
-
-
