+++ src/components/organisms/RichMenu/RichMenuDescriptionContent/index.ts
@@ -1,0 +1,1 @@
+export * from './RichMenuDescriptionContent';
