+++ src/components/organisms/RichMenu/RichMenuDescriptionContent/type.ts
@@ -1,0 +1,2 @@
+export type RichMenuDescriptionContentProps = {
+}
