+++ src/components/organisms/RichMenu/RichMenuListContent/index.ts
@@ -1,0 +1,1 @@
+export * from './RichMenuListContent';
