+++ src/components/organisms/RichMenu/RichMenuListContent/type.ts
@@ -1,0 +1,2 @@
+export type RichMenuListContentProps = {
+}
